Preparing codebase to enable it to receive any message types
This is the case for SUBSCRIBE and PUBLISH: either one of them can get
a PUBACK, PUBREL, SUBACK or PUBLISH back, in any order.

This changes the way this API behaves quite a lot, because an object
isn't specifically bound anymore to another, but they can all call each
other up.

This commit is the preparation for that.